Move Swatches and ColorSwatch into their own file
